2017 - Dec 23
Interpret the program in part 1, sum up the mul calls.
Figure out what it does in part 2, write as code, simplify and simplify the code. When the algorithm is clear, it is possible to write a Eratosthenes Sieve implementation that can be used to find the non-primes in the range with given step.